requirements.txt
# Requirements
scikit-image
Flask
numpy
matplotlib
scipy
fask-restful
Pillow
pixelsort
